+++ lisp/test-files/range-restriction-and-early-termination.c
@@ -1,0 +1,1 @@
+DEFUN ("safe-length", Fsafe_length, Ssafe_length, 1, 1, 0)